refactor(svelte): internalize plot interaction assembly
