+++ .changeset/internalize-plot-interaction-assembly.md
@@ -1,0 +1,8 @@
+---
+"@ggsvelte/svelte": patch
+---
+
+# Internalize plot interaction assembly
+
+Internalize controller construction, deferred sibling wiring, shared Candidate
+projection, and lifecycle registration in a deep plot-interaction assembly module.
